refactor: update chat and user handling to enhance privacy and functionality
- Replaced email exposure with display names in chat messages and user search results.
- Updated Socket.io event payloads to remove email fields and include sender names.
- Modified user search to return follow status for authenticated users.
- Adjusted README and OpenAPI specifications to reflect changes in data handling.
