+++ apps/frontend/index.html
@@ -2,7 +2,7 @@
 <html lang="en" class="dark">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>%VITE_APP_NAME%</title>
     <meta
@@ -21,7 +21,7 @@
       content="CineConnect helps you discover trending movies, track your watchlist, and chat with the movie community in real time."
     />
     <meta property="og:url" content="http://localhost:8080/" />
-    <meta property="og:image" content="/vite.svg" />
+    <meta property="og:image" content="/favicon.svg" />
 
     <meta name="twitter:card" content="summary_large_image" />
     <meta name="twitter:title" content="%VITE_APP_NAME%" />
@@ -29,7 +29,7 @@
       name="twitter:description"
       content="CineConnect helps you discover trending movies, track your watchlist, and chat with the movie community in real time."
     />
-    <meta name="twitter:image" content="/vite.svg" />
+    <meta name="twitter:image" content="/favicon.svg" />
   </head>
   <body>
     <div id="root"></div>
